+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.0] - 2025-01-06
+
+### Added
+
+- Add `enable` settings to enable or disable the extension
+
+### Fixed
+
+- Fix typos the file generation process
+
 ## [2.8.0] - 2024-12-23
 
 ### Added
@@ -228,7 +238,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-nestjs-generator/compare/v2.8.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-nestjs-generator/compare/v2.9.0...HEAD
+[2.9.0]: https://github.com/ManuelGil/vscode-nestjs-generator/compare/v2.8.0...v2.9.0
 [2.8.0]: https://github.com/ManuelGil/vscode-nestjs-generator/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/ManuelGil/vscode-nestjs-generator/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/ManuelGil/vscode-nestjs-generator/compare/v2.5.0...v2.6.0